+++ src/test/app/agent/blog_agent_test.py
@@ -16,7 +16,7 @@ class BlogAgentTest(unittest.TestCase):
         agent = TestBlogAgent.of_config(
             agent_name,
             run_context.get_app_config().to_dict(),
-            load_agent_config(agent_name, False))
+            load_agent_config(agent_name, True))
 
         result = agent.run(run_context)